requirements.txt
python-dotenv
sec-api
requests
beautifulsoup4 